% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{AIC.RegressionByComposition}
\alias{AIC.RegressionByComposition}
\title{Compute Akaike Information Criterion from a regression by composition}
\usage{
\method{AIC}{RegressionByComposition}(object, ..., k = 2)
}
\arguments{
\item{object}{a RegressionByComposition object;
usually the result of a call to rbc()}

\item{...}{ignored}

\item{k}{numeric, the \emph{penalty} per parameter to be used; ‘k = 2’ is the classical AIC.}
}
\description{
Compute Akaike Information Criterion from a regression by composition
}
