% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Bernoulli}
\alias{Bernoulli}
\title{Bernoulli distribution as a CompositeFamily}
\usage{
Bernoulli(prob = 0.5)
}
\arguments{
\item{prob}{the probability of a success}
}
\value{
a new BinaryFamily object
}
\description{
Bernoulli distribution as a CompositeFamily
}
\examples{
dist <- Bernoulli()
dist$probability()
}
