% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{coef.RegressionByComposition}
\alias{coef.RegressionByComposition}
\title{Extract regression coefficients from a regression by composition}
\usage{
\method{coef}{RegressionByComposition}(object, ...)
}
\arguments{
\item{object}{a RegressionByComposition object;
usually the result of a call to rbc()}

\item{...}{ignored}
}
\description{
Extract regression coefficients from a regression by composition
}
