% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{fitted.RegressionByComposition}
\alias{fitted.RegressionByComposition}
\title{Compute fitted values from a regression by composition}
\usage{
\method{fitted}{RegressionByComposition}(object, ...)
}
\arguments{
\item{object}{a RegressionByComposition object;
usually the result of a call to rbc()}

\item{...}{further arguments passed to the R6 method $fitted() associated with the model's CompositeFamily}
}
\description{
Compute fitted values from a regression by composition
}
