% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{logLik.RegressionByComposition}
\alias{logLik.RegressionByComposition}
\title{Extract log-likelihood from a regression by composition}
\usage{
\method{logLik}{RegressionByComposition}(object, ...)
}
\arguments{
\item{object}{a RegressionByComposition object;
usually the result of a call to rbc()}

\item{...}{ignored}
}
\description{
Extract log-likelihood from a regression by composition
}
