% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{LogNormal}
\alias{LogNormal}
\title{Lognormal distribution as a CompositeFamily}
\usage{
LogNormal(meanlog = 0, sdlog = 1)
}
\arguments{
\item{meanlog}{the mean of the logarithm}

\item{sdlog}{the standard deviation of the logarithm}
}
\value{
a new ContinuousFamily object
}
\description{
Lognormal distribution as a CompositeFamily
}
\examples{
dist <- LogNormal()
log(dist$quantile(0.95))
}
