% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{Moebius}
\alias{Moebius}
\title{Moebius flow}
\format{
An object of class \code{Flow} (inherits from \code{R6}) of length 6.
}
\usage{
Moebius
}
\description{
Moebius flow
}
\keyword{datasets}
