% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{Normal}
\alias{Normal}
\title{Normal distribution as a CompositeFamily}
\usage{
Normal(mean = 0, sd = 1)
}
\arguments{
\item{mean}{the mean}

\item{sd}{the standard deviation}
}
\value{
a new ContinuousFamily object
}
\description{
Normal distribution as a CompositeFamily
}
\examples{
dist <- Normal()
dist$quantile(0.95)
}
