% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{predict.RegressionByComposition}
\alias{predict.RegressionByComposition}
\title{Compute predicted values from a regression by composition}
\usage{
\method{predict}{RegressionByComposition}(object, newdata, ...)
}
\arguments{
\item{object}{a RegressionByComposition object;
usually the result of a call to rbc()}

\item{newdata}{data.frame containing new data}

\item{...}{further arguments passed to the R6 method $fitted() associated with the model's CompositeFamily}
}
\description{
Compute predicted values from a regression by composition
}
