% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{residuals.RegressionByComposition}
\alias{residuals.RegressionByComposition}
\title{Compute 'residuals' from a regression by composition}
\usage{
\method{residuals}{RegressionByComposition}(object, ...)
}
\arguments{
\item{object}{a RegressionByComposition object;
usually the result of a call to rbc()}

\item{...}{ignored}
}
\value{
a vector of probabilities of the same length as the data
}
\description{
Compute 'residuals' from a regression by composition
}
