% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{ScaleOdds}
\alias{ScaleOdds}
\title{ScaleOdds flow}
\format{
An object of class \code{Flow} (inherits from \code{R6}) of length 6.
}
\usage{
ScaleOdds
}
\description{
ScaleOdds flow
}
\keyword{datasets}
