% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{ScaleRisk0}
\alias{ScaleRisk0}
\title{ScaleRisk0 flow}
\format{
An object of class \code{Flow} (inherits from \code{R6}) of length 6.
}
\usage{
ScaleRisk0
}
\description{
ScaleRisk0 flow
}
\keyword{datasets}
