% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starr.R
\docType{data}
\name{starr}
\alias{starr}
\title{Growth from birth to 3 years in healthy babies in the US}
\format{
\subsection{\code{starr}}{

A data frame with 104,798 rows and 5 columns:
\describe{
\item{id}{Anonymized identifier}
\item{sex}{Sex of baby}
\item{age}{Age of baby, in days}
\item{height}{Jittered height of baby, in cm}
\item{weight}{Jittered weight of baby, in kg}
...
}
}
}
\source{
\doi{doi:10.5061/dryad.4j0zpc8jf}
}
\usage{
starr
}
\description{
Growth from birth to 3 years in healthy babies in the US
}
\references{
\doi{doi:10.1186/s12874-024-02145-1}
}
\keyword{datasets}
