% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{summary.RegressionByComposition}
\alias{summary.RegressionByComposition}
\title{Summary of a regression by composition}
\usage{
\method{summary}{RegressionByComposition}(object, compact = FALSE, ...)
}
\arguments{
\item{object}{a RegressionByComposition object;
usually the result of a call to rbc()}

\item{compact}{logical; should coefficients from all flows be compressed into a single matrix?}

\item{...}{ignored}
}
\description{
Summary of a regression by composition
}
