% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{Translate}
\alias{Translate}
\title{Translate flow}
\format{
An object of class \code{Flow} (inherits from \code{R6}) of length 6.
}
\usage{
Translate
}
\description{
Translate flow
}
\keyword{datasets}
