% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\docType{data}
\name{TranslateRisk1}
\alias{TranslateRisk1}
\title{TranslateRisk1 flow}
\format{
An object of class \code{Flow} (inherits from \code{R6}) of length 6.
}
\usage{
TranslateRisk1
}
\description{
TranslateRisk1 flow
}
\keyword{datasets}
