% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{vcov.RegressionByComposition}
\alias{vcov.RegressionByComposition}
\title{Extract variance-covariance matrix from a regression by composition}
\usage{
\method{vcov}{RegressionByComposition}(object, ...)
}
\arguments{
\item{object}{a RegressionByComposition object;
usually the result of a call to rbc()}

\item{...}{ignored}
}
\description{
Extract variance-covariance matrix from a regression by composition
}
